added a bunch of stuff (see descr.)
i was kinda bored this morning.. added some comments to the code,
messed around with the particle system a bit, so it now shoots out a
bunch of red cubes from the middle. It’s pretty laggy though so we need
to do some performance optimizations. we also need to figure out how we
want to do the water stuff, looks weird now